requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
tavily-python>=0.3.0
rich>=13.7.0
python-dotenv>=1.0.0
gradio>=5.0.0
plotly>=5.18.0
